requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
crispy-bootstrap5==0.6
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-braces==1.15.0
django-cloudinary-storage==0.3.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==21.1
django-jquery==3.1.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
requests==2.27.1
self==2020.12.3
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
